🔄 refactor(orderflow): simplify drawing logic by extracting circle drawing to a separate function
